Fix a DCHECK hit when serializing the 'offset' shorthand
The serialization code for the 'offset' shorthand assumes that when the
'offset-direction' and 'offset-rotate' longhands have non-initial values
only if the 'offset-path' longhand has a non-initial value, which is not
true.

This patch handles the case by making the shorthand serialization fail,
so it falls back to serialize the individual longhands.

Fixed: 1292163
Change-Id: Ibd341364cb52fb1c4e6335337a15e12ed42011da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3441289
Auto-Submit: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967762}