[WebAuthn] Allow autofill to asynchronously retrieve WebAuthn creds
Autofill can offer to provide Web Authentication credentials alongside
username/passwords for login fields on desktop.

In order for this to work on Android, the credentials have to be
fetched over an asynchronous GMS Core API.

This change initiates the credential retrieval at the same time that
server suggestions are being downloaded, and adds a callback for when
the retrieval is complete.

Bug: 1171985
Change-Id: I8227aaf8b8ef166b47f2ed164babcf3d0a33f5f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3402849
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967764}