CSS parsing utils preparations for font-palette-values parsing
Add a method to check for dashed identifiers "--myRuleName",
add an option to exclude system colors from ConsumeColor parsing.

With that, rename IsSystemColor() to IsDeprecatedSystemColor()
and add a new method IsSystemColor() to StyleColor to account
for changes in color names in the spec. [1][2]

[1] https://www.w3.org/TR/css-color-4/#css-system-colors
[2] https://www.w3.org/TR/css-color-4/#deprecated-system-colors

Bug: 1170794
Change-Id: Idff54a25381c4ff6395c30aff6fe383984ceadce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3443344
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967802}