gpu: Refactor vulkan init
Clean up after r966326.

Refactor VulkanFunctionPointers so the library and lock are private, and
combine setting the library with BindUnassociatedFunctionPointers so
that lock can be made private. This means splitting Bind into two
versions, one taking a library, and one taking a GetProcAddr functor.

Make VulkanInstance::Initialize take a path to the library, and only
load the library when needed. Also only unload the library if it's
loaded by Initialize.

Update all callers of VulkanInstance::Initialize to pass in the library
path instead of loading the library directly.

Bug: 1293005
Change-Id: Id11a3e50ea408f3e930c1b63baa0d7c145bd29a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3441274
Reviewed-by: Peng Huang <penghuang@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967824}