Add tests for sandboxed iframes and null initiator.
There weren't any tests that asserted that the initiator was null when
a WebRequest is observed inside a sandboxed iframe.

Sandboxed iframe requests are considered 3rd party as well because of
the null initiator, so add a test for declarative net request for that.

Change-Id: I128335fc79a0c5c9c9d9c8847d7770a9e4fac6b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3437326
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967833}