Add HttpNetworkSessionParams.ignore_ip_address_changes
When ignore_ip_address_changes = true, objects used by a
HttpNetworkTransaction are asked not to perform disruptive work after
there has been an IP address change (which usually means that the
"default network" has possibly changed). ignore_ip_address_changes
defaults to false, in which case the behavior is the same as before.

Note: ignore_ip_address_changes will be set to false for
HttpNetworkSessions that are bound to a specific network. In this case,
we know that the session will continue using the network it is bound to
even after an IP address change. Hence, the usually performed disruptive
work (e.g., emptying socket pools) should not be performed.

Bug: 1292548
Change-Id: I3e68b5bc20072b7465922afe2aa7bde78a2e2afe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3427689
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Stefano Duo <stefanoduo@google.com>
Cr-Commit-Position: refs/heads/main@{#967854}