Make HideHelper set the compositor for animations so they stop ticking.
This fixes most of the remaining test failures with
https://crrev.com/c/3200347 (which in turn blocks a bunch of CLs for
https://crbug.com/696374 which in turn blocks shipping independent
transform properties).

This is needed because, with the above CL, cc is stricter about
animations being unregistered if they animate an element whose element
ID is no longer in the layer tree.

I took this approach of adding a little bit more of what SetRootLayer
does into this code, rather than the approach of simplifying the code
and calling ui::Compositor::SetRootLayer, because the alternative caused
a decent number of test failures, and this option just worked.

Bug: 900241
Change-Id: I58f309e1a9fe84d308c937701eae9e6ea66cd499
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3440535
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967861}