[FLEDGE] Introduce Bid and ScoredBid to AuctionRunner.
Currently, only interest groups create bids and each bid is only scored
once, so it's simplest to have each bidder own its score and bid, and
to use the BidderWorkletBidPtr mojo class to store each bid, however,
with component auctions, this will change.

Introduce a Bid class that will down the line be able to either
reference a source bidder or seller who made the bid, and allow
component seller to copy a bid but introduce their own bid value.

Also introduce a ScoredBid class, that will be able to score data
(desireability score, data version) for each seller worklet that's
scored a bid.

This CL also reduces memory usage slightly, as the new Bid struct is
thrown out as soon as the bid has lost the auction, since it's no
longer needed, and in the new model, there's no place to store it.

Bug: 1288865
Change-Id: Ie042b9ef6d372b6910507332610d507ce6d99e88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3442814
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967865}