[Color Pipeline] Replace GetThemeProviderForProfile with widget based ThemeProvider in ChromeCustomizeThemesHandler.
VerifyThemeColors only tests that the developer called the color APIs
correctly. It calls the same function (GetAutogeneratedThemeColors) that
is called in ChromeCustomizeThemesHandler::GetChromeThemes or later down
the line in BrowserThemePack for ThemeProvider colors and simply
compares the results of effectively the same function call. On moving to
a widget based ThemeProvider we will not have a BrowserWindow in this
test setting and therefore will not have a ThemeProvider to test the
specific colors generated. Therefore this CL removes VerifyThemeColors
from the unit tests.

We will be able to remove the logic that returns a placeholder color
when we switch to ColorProvider since we will be guaranteed a
ColorProvider for WebContents.

Bug: 1056758
Change-Id: I2c761d04e89e26416069324e06796e4802847430
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3435147
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Elaine Chien <elainechien@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967894}