personalization: Add Google Photos support to fetch a single photo
This CL adds an item ID parameter to the Wallpaper app's existing mojo
API for fetching photos from Google Photos. If provided, the ID
specifies one photo to fetch. This option takes precedence over the
option to fetch an album if an album ID is provided. If neither is
provided, the whole library is fetched.

We will need to fetch a single photo from Google Photos in order to set
it as a device's wallpaper.

Fixed: b/217612619
Change-Id: I1cccafbc8e0a2559b3985aae9310744b36a6ba53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3433993
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Commit-Queue: Colin Kincaid <ckincaid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967901}