Add app list reorder nudge view
This cl creates the AppListNudgeController class to controls if a toast
like nudge should be shown in the app list. The AppListToastView is
used to implement the nudge view.
Also rewrite the reorder undo container view class to toast container
view class. This would allow the nudge to use the same container view
with the reorder undo toast.

This cl is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/3388329.
Design comments can be found in the previous cl.

Bug: 1277668
Change-Id: Ic5346e121f7abd65969ce8a405af0b6bb3f488ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3441110
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967906}