[journeys] Caps the number of related searches to 5
Also further simplifies the code by removing a set that keeps track
of unique related searches in order to improve the code readability.
The number of related searches is so small that looking for duplicates
in a vector will not cause any measurable performance regression.

Fixed: 1294369
Change-Id: I30ff700c57ce504a6bc26a53742213ab2b7a52d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3441089
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Mohamad Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967928}