Move wallpaper cropping to a general utility function.
ash/ambient is going to depend on this for the animated screensaver.
The UX requirement for handling photos was to replicate what we do
for wallpaper. So all this CL does is move the existing wallpaper
cropping to a common location where ash/ambient can use it as well.
The "scaling" portion is intentionally not copied since ash/ambient
will not reuse that part.

Bug: b:217271214
Cq-Include-Trybots: luci.chrome.try:linux-chromeos-chrome
Change-Id: I9f39275dd7923978b077ea7549dd90ffb6cee355
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3439033
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#967945}