Clean up promises in the color enhancer extension storage class
As part of refactoring the storage class of the color enhancer
extension, it has transitioned from fetching the values on demand from
the chrome.storage API to caching the values and listening for changes.
This is a cleanup CL to remove promises that return immediately from
the storage class's interface.

This change is a pure refactor.

AX-Relnotes: n/a.
Bug: 1289423
Test: Manually inspected extension behavior.
Change-Id: If1c877136bafc1f28ded742e464ebb7dc4046783
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438785
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Josiah Krutz <josiahk@google.com>
Cr-Commit-Position: refs/heads/main@{#967948}