Allow attribution simulator to set noise mode and report mode
We replace AttributionStorageDelegateImpl's single `bool debug_mode_`
field with separate enums, but still set both of them based on the
kConversionsDebugMode command-line switch in the real browser.

We will add a seeded-RNG mode in a followup CL, as there are more
complex design questions to resolve for that feature.

Bug: 1294272
Change-Id: I280ab3510a748b02075664b57690268ba7cd0a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3439382
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967952}