capture_selfie_cam: Camera disconnection
This CL adds the logic to handle the case when a selected camera
gets disconnected. We give it a grace period so that it may
reconnect again (which should help in the case of a flaky camera).
After the grace period expires, the selected camera ID is cleared.

Fixed: 1293615
Test: Added new tests.
Change-Id: I77cece84ef652d330eb91f9481aa07308546e44a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438027
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967953}