Reland: Restore original parents of aria-owns owned children when owner is removed
When aria-owns relationship is set, the owned children accessible
parent is set to the owner. But, if the owner is removed, the
relationship should be reset to the original parent.

This issue could be present for a long time, though we detected at
least the problem happens more frequently after the change
https://chromium.googlesource.com/chromium/src/+/749ab8571dc14a1627e22a61c70974ab1995046d
This is because before, invalidation would cover a bigger part of the
tree.

Originally reverted because of crbug.com/1289625 issue detected with
clusterfuzz. The problem was with trying to reattach parent even when
the accessible tree was disposed. Now the RestoreParentOrPrune method
will not assert if the parent is not available. We also avoid
reparenting if the object cache has been disposed.

We also add an extra test for the case the ARIA owner is removed, but
there is still another ARIA owner.

An specific test for clusterfuzz scenario is not needed as it is
already covered by the new test (disposal of the accessible tree on
test shutdown when there is still an ARIA owner).

Bug: 1287181
Change-Id: If71451a42b2d077f88f590d92096711000dd94f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3419685
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Cr-Commit-Position: refs/heads/main@{#967957}