[iOS] Make method a shared version in //components/safe_browsing
GetEventTypeFromThreatSource is now a method used in both non-iOS and
iOS versions of SafeBrowsingBlockingPage. To reflect this change, the
method is being moved to SafeBrowsingMetricsCollector as the method is
only really used with
SafeBrowsingMetricsCollector::AddSafeBrowsingEventToPref() and it
returns an enum which is created in the class as well.

Bug: 1291674
Change-Id: I578db12cb59ce2952ed80438a9c23afd06ff31b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3440735
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#968009}