Special case APP_ID sources for UKM Source Pruning.
This CL has made this section quite a mess, so sorry for this. I'm not 100% sure if it makes sense to leave the seen/unseen experiment. I decided to leave it running and we can compare however we want. Eventually we will clean this up much better.

The testing isn't really exhaustive either unfortunately, but I'm not sure it's worth complex testing of the pairwise experiments if we're not keeping them for long.

Bug: 1268552
Change-Id: Icfcc1aa115affef6fdef44c1ce8335d211e81fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3433853
Reviewed-by: Yue Ru Sun <yrsun@chromium.org>
Auto-Submit: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#968033}