[OnDemand AT] Add histograms to track interactions of OnDemand and ComputeAXMode Android
This CL adds a new histogram to the Android accessibility code that
will allow us to track the effect of both the OnDemand AT and the
ComputeAXMode features in unison to help determine whether or not new
AXModes would be beneifical for users.

We include new histograms, code to track them, and a series of tests
to check that the histograms are being recorded as expected.

AX-Relnotes: N/A
Bug: 1227192
Change-Id: I770a359f70eb929867195e91d7c8ec4bdd9091da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438414
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#968075}