Rename and clean up usages for LocalFrameView::scrollable_areas_
After https://crrev.com/c/3424999, `LocalFrameView::scrollable_areas_`
contains scrollable areas which either a) are user-scrollable, visible
to hit testing, and do in fact overflow, or b) overflow in the block
flow direction. This is confusing, and it is a result of
`scrollable_areas_` being used both for scroll anchoring and for cases
which should use `user_scrollable_areas_` (itself introduced in
https://crrev.com/c/3306960).

This CL changes every usage of `scrollable_areas_` unrelated to scroll
anchoring to use `user_scrollable_areas_`, and renames
`scrollable_areas_` to `scroll_anchoring_scrollable_areas_` to reflect
its current usage. It also splits the code in
`PaintLayerScrollableArea::UpdateScrollableAreaSet` to use different
conditions for adding to and removing from each of the two sets.

Bug: 1293192
Change-Id: I47d1fdeb8e9f1a4cc534ce473225486caf493453
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3427547
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#968097}