Block sync install for default Chrome apps that have migrated to web apps
This CL adds a step in the extension sync installation pipeline to
prevent the installation of default Chrome apps that are intended to be
migrated to an equivalent web app.

This solves an app duplication bug in the default web app migration
where users with the old Chrome app still in their sync profile would
see the Chrome app reinstall alongside the web app on the next start up.

Bug: 1290718
Change-Id: Ie5ff48ba0eeaa8bb231c3cf89148a7508f5708d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3340991
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Cr-Commit-Position: refs/heads/main@{#968125}