[Extensions] Remove 'Options' entry from context menu if no options page
The 'options' entry in the context menu will only be created if the
extension has an options page. This differs from the previous
functionality, which added an 'options' entry for all non component
extensions, and then disabled it if there was no options page.

To avoid checking for visibility in two places, moved all visibility
computations to `InitMenu()`. This doesn't change functionality, and
reduced potential confusion in tests.


Bug: 1289996
Change-Id: I4f01780b66b46b46465d9fcc586ac8f6e3c270e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3425463
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#968126}