Arc: Always restore a fullscreen ARC window to kNormal window state.
Before the window state restore history changes [1][2], when restoring
a app-requested ARC window fullscreen, it will always changes back to
kNormal window state. (Note this restoring-from-fullscreen behavior is
different with the system-requested ARC window fullscreen, in which the
window will return back to its pre-fullscreen window state, which can be
non-kNormal window state). The [1][2] CLs changed the fullscreen restore
behavior so that for app-requested ARC window, when exiting fullscreen,
it also changes back to its pre-fullscreen window state. However, due to
restrictions on ARC side, the ARC window is not allowed to change to
kMaximized window state when it's in fullscreen, thus causing this
regression.

This CL temporarily reverted the fullscreen restore behavior so that
a fullscreen ARC window will always restore to kNormal window state. The
CL is meant to merge back to M99.
After M99, we'll try to find a long-term solution to see if it's
possible to unify the exiting fullscreen behavior for the two fullscreen
states for ARC window, so that its behavior is also consistent with
that of Chrome windows.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3314057
[2] https://chromium-review.googlesource.com/c/chromium/src/+/3346502

Bug: b/215710461
Change-Id: I585c6b603b41bfd913eca8070196ee28000ece85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438399
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Kazuki Takise <takise@chromium.org>
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#968136}